Sample queries

+++ BandwidthMonitorService.Domain/Models/Sample.cs
@@ -11,9 +11,9 @@ namespace BandwidthMonitorService.Domain.Models
         public string Id { get; set; }
         public int Timestamp { get; set; }
         public string Url { get; set; }
-        public long BytesRead { get; set; }
-        public long TotalReads { get; set; }
+        public double BytesRead { get; set; }
+        public double TotalReads { get; set; }
         public TimeSpan Elapsed { get; set; }
-        public long RoundTripTime { get; set; }
+        public double RoundTripTime { get; set; }
     }
 }
